Changed example and some optimizations
Now you can navigate throgh the different adapter examples and have a comparison between the old and new way in the code.

Also removing the inline to certain functions the bytecode is significantly smaller, now it's just a bit larget compared to a regular adapter.

There's a new extension function to allow better performance items